extract timestamp into a separate file
